chore: move styled system inside src
