+++ client/.gitignore
@@ -20,5 +20,5 @@ coverage
 **/locales/**/*.ts
 
 # Generated by Panda CSS
-styled-system/generated
-styled-system/.env
+src/styled-system/generated
+src/styled-system/.env
